added the missing files for the cms package. working on the cms package Implemented the Menu functionality, and now i can use it on the entire page. But. there are still some bugs i need to fix

+++ src/controllers/cmsControllers.py
@@ -4,14 +4,9 @@ import models.cmsModels as cms
 from controllers.BaseControllers import LoginController
 from lib.halicea.HalRequestHandler import HalRequestHandler as hrh
 from forms.cmsForms import CMSContentForm
-from google.appengine.api import memcache
 from google.appengine.ext import db
 from lib import messages
 from django.utils import simplejson
-from models.cmsModels import Comment, ContentTag
-from forms.cmsForms import CommentForm
-import logging
-
 
 class CMSBaseController(hrh):
     def __init__(self, *args, **kwargs):
@@ -26,16 +21,22 @@ class CMSLinksController(CMSBaseController):
     def SetOperations(self):pass
     @AdminOnly()
     @View(templateName='CMSLinks.html')
-    def index(self, *args):
+    def index(self, menu='cms',*args):
         limit = 100
         offset = 0
         try:
             offset = int(self.params.offset)
         except:
             pass
-        cmsLinks = cms.CMSLink.GetLinkTree()
+        resultMenu = cms.Menu.get_by_key_name(menu)
+        if not resultMenu:
+            root = cms.CMSLink.get_or_insert(
+                         key_name = menu, 
+                         AddressName=menu, Name=menu, 
+                         Creator=self.User, ContentTypeNumber=2)
+            resultMenu = cms.Menu.CreateFromLinkRoot(menu, 'None', None, root, True) 
         contents = cms.CMSContent.all().order('-DateCreated').fetch(limit=limit, offset=offset)
-        return {'cmsLinks':cmsLinks, 'contents':contents}
+        return {'menu':resultMenu, 'contents':contents}
     @AdminOnly()
     def save(self, *args):
         addressName = self.params.addressName
@@ -46,32 +47,53 @@ class CMSLinksController(CMSBaseController):
         else:
             parent = None
         order= int(self.g('order'))
-        content=self.params.content
+        content=self.params.content or None
+        contentType = cms.ContentType.StaticPage
         if content:
             content = cms.CMSContent.get(content)
+            contentType=cms.ContentType.CMSPage
+        if not self.params.HasContent:
+            contentType = cms.ContentType.NoContent
         creator= self.User
         if True: #TODO: validation
-            cms.CMSLink.CreateNew2(addressName, name, parent, order, content, creator, _isAutoInsert=True)
-
+            cms.CMSLink.CreateNew2(addressName, name, parent, order, content, contentType, creator, _isAutoInsert=True)
+        
         if self.isAjax:
-            return "Links Created"
+            return self.LinksTree(self.params.menu)
         else:
             return self.index()
     @AdminOnly()
     def delete(self, *args):
         lnk=cms.CMSLink.get(self.params.key)
         if lnk:
-            if lnk.Content.content_cms_links.count()==1:
+            if lnk.Content and lnk.Content.content_cms_links.count()==1:
                 lnk.Content.delete()
             lnk.delete()
             self.status='Link is deleted'
-            self.redirect(self.get_url())
         else:
             self.status="Link is invalid";
-            self.redirect(self.get_url())
-    def LinksTree(self):
-        return {'cmsLinks':cms.CMSLink.GetLinkTree()}
+        return self.isAjax and self.status or self.redirect(self.get_url())
 
+    def LinksTree(self, menu):
+        return cms.Menu.get_by_key_name(menu).to_list()
+    
+class MenuController(CMSBaseController):
+    def __init__(self, *args, **kwargs):
+        super(MenuController, self).__init__(*args, **kwargs)
+    @ClearDefaults()
+    @Default('view')
+    def SetOperations(self):pass
+    
+    @CachedResource()
+    def view(self, menu='cms'): 
+        result=  cms.Menu.get_by_key_name(menu)
+        if result:
+            return result.to_list(self.params)
+        else:
+            return "No Menu Found"
+    def index(self):
+        return {'menus':cms.Menu.all()}
+    
 class CMSContentController(CMSBaseController):
     def __init__(self, *args, **kwargs):
         super(CMSContentController, self).__init__(*args, **kwargs)
